requirements.txt
markdown2
weasyprint
html2text
imgkit
jinja2
bs4
fastapi
unvicorn
pyjwt